refactor: 방 생성 및 선택 로직 별도 커스텀 훅으로 분리
